refactor(metadata): 例外 import を canonical owner へ移行する (#3958)
